add project_age and issue_id to analytics
